requirements.txt
click>=6.6
Flask>=0.11.1
tellcore-py>=1.1.3
plotly>=1.12.12
